@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -29,43 +29,67 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.3
-  flutter_spinkit: ^5.0.0
-  shared_preferences: ^2.0.6
-  camera: ^0.8.1+3
-  qr_code_scanner: ^0.5.1
-  pinput: ^1.2.0
-  intl: ^0.17.0
-  flutter_svg: ^0.22.0
-  fluttertoast: ^8.0.7
-  provider: ^5.0.0
-  permission_handler: ^8.1.1
-  flutter_local_notifications: ^6.0.0
-  app_settings: ^4.1.1
-  adaptive_dialog: ^1.0.1
-  path_provider: ^2.0.2
-  http: ^0.13.3
-  json_annotation: ^3.0.1
-  unique_identifier: ^0.2.2
-  upgrader: ^3.3.1
-  package_info: ^2.0.2
-  modal_progress_hud: ^0.1.3
-  pushy_flutter: ^2.0.3
-  get_it: ^7.2.0
-  image_picker: ^0.8.4+5
-  fbroadcast: ^1.1.1
-  smart_select: ^4.3.2
-  geolocator: ^7.4.0
-  latlng: ^0.1.0
-  signature: ^4.1.1
-  pin_code_fields: ^7.3.0
+  # Core UI
+  cupertino_icons: ^1.0.6
+  flutter_spinkit: ^5.2.0
+  flutter_svg: ^2.0.9
 
+  toastification: ^2.1.0
+  modal_progress_hud_nsn: ^0.5.1  # Replacement for modal_progress_hud
+  
+  # Storage
+  shared_preferences: ^2.2.2
+  path_provider: ^2.1.2
+  
+  # Camera & QR
+  camera: ^0.10.5+9
+  mobile_scanner: ^5.0.0  # Replace qr_code_scanner
+  image_picker: ^1.0.7
+  
+  # Input
+  pinput: ^3.0.1
+  pin_code_fields: ^8.0.1
+  signature: ^5.4.1
+  
+  # Localization
+  intl: 0.20.2
+  
+  # State Management
+  provider: ^6.1.1
+  get_it: ^7.6.7
+  
+  # Notifications & App Updates
+  flutter_local_notifications: ^17.2.2
+  app_settings: ^5.1.1
+  upgrader: ^8.4.0
+  package_info_plus: ^4.2.0  # Replacement for package_info
+  pushy_flutter: ^2.0.16
+  
+  # Utility
+  permission_handler: ^11.2.0
+  http: ^1.2.0
+  json_annotation: ^4.8.1
+  device_info_plus: ^10.1.2
+  adaptive_dialog: ^1.10.1
+  fbroadcast: ^2.0.0
+  smart_select: ^4.3.2
+  
+  # Location
+  geolocator: ^10.1.0
+  latlng: ^1.0.0
+  choice: ^2.3.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  json_serializable: ^3.3.0
-  build_runner: ^1.10.9
+  json_serializable: ^6.7.1
+  build_runner: ^2.4.7
+  watcher: ^1.1.0
+  pub_semver: ^2.1.4
+
+dependency_overrides:
+  win32: ^5.12.0
+  file: ^7.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -83,6 +107,7 @@ flutter:
     - assets/locale/en.json
     - assets/locale/he.json
     - assets/locale/ru.json
+    - assets/locale/ar.json
 
   # To add assets to your application, add an assets section, like this:
   # assets:
