@@ -19,7 +19,7 @@ import 'package:goport/Models/Truck.dart';
 import 'package:goport/Network/GoPortApi.dart';
 import 'package:goport/Providers/GeneralProvider.dart';
 import 'package:goport/Providers/LocationProvider.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
+import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -43,8 +43,8 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
   List<Event> _events = [];
   final _cardHeight = 140.0;
   bool _enterTrackVisible = true;
-  AnimationController animation;
-  Animation<double> _fadeInFadeOut;
+  late AnimationController animation;
+  late Animation<double> _fadeInFadeOut;
 
   @override
   void initState() {
@@ -79,7 +79,8 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
-    final generalProvider = Provider.of<GeneralProvider>(context, listen: false);
+    final generalProvider =
+        Provider.of<GeneralProvider>(context, listen: false);
     if (state == AppLifecycleState.resumed) {
       generalProvider.setShowBackButton(false);
       _checkExistsDraft();
@@ -97,8 +98,8 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
   _checkExistsDraft() async {
     final generalProvider =
         Provider.of<GeneralProvider>(context, listen: false);
-    bool res =
-        await GoPortApi.instance.checkExistsDraft(generalProvider.driver.tz);
+    bool res = await GoPortApi.instance
+        .checkExistsDraft(generalProvider.driver!.tz ?? "");
     if (res) {
       setState(() {
         _hasJobDraft = true;
@@ -110,12 +111,14 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
     final generalProvider =
         Provider.of<GeneralProvider>(context, listen: false);
     String guidID = await GoPortApi.instance
-        .getJobCardGuidIDByDriver(generalProvider.driver.tz);
+            .getJobCardGuidIDByDriver(generalProvider.driver?.tz ?? "") ??
+        "";
     if (guidID != null && guidID != "null") {
       guidID = guidID.replaceAll("\"", "");
       generalProvider.setShowBackButton(true);
       WidgetsBinding.instance.removeObserver(this);
-      await Navigator.of(context).pushNamed("JobCardScreen", arguments: {"guidID": guidID});
+      await Navigator.of(context)
+          .pushNamed("JobCardScreen", arguments: {"guidID": guidID});
       WidgetsBinding.instance.addObserver(this);
       generalProvider.setShowBackButton(false);
       _checkJobCardExists();
@@ -127,13 +130,16 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
         Provider.of<GeneralProvider>(context, listen: false);
     final driver = generalProvider.driver;
     final SharedPreferences prefs = await _prefs;
-    Truck truck;
-    Truck trailer;
-    if (prefs.getString(Const.prefsTruck)!= null) {
-      truck = Truck.fromJson(jsonDecode(prefs.getString(Const.prefsTruck)));
+    Truck? truck;
+    Truck? trailer;
+    if (prefs.getString(Const.prefsTruck) != null) {
+      truck =
+          Truck.fromJson(jsonDecode(prefs.getString(Const.prefsTruck) ?? ""));
     }
-    if (prefs.getString(Const.prefsTrailer)!= null) {
-      trailer =  Truck.fromJson(jsonDecode(prefs.getString(Const.prefsTrailer)));
+    if (prefs.getString(Const.prefsTrailer) != null &&
+        prefs.getString(Const.prefsTrailer) != "null") {
+      trailer =
+          Truck.fromJson(jsonDecode(prefs.getString(Const.prefsTrailer) ?? ""));
     }
 
     String fullName = "";
@@ -149,9 +155,11 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
 
     if (truck != null && trailer != null) {
       _getVehicleDetails(truck.num, trailer.num);
+    } else if (truck != null) {
+      _getVehicleDetails(truck.num, "");
     }
 
-    EventResponse res = await GoPortApi.instance.getEvents();
+    EventResponse? res = await GoPortApi.instance.getEvents();
     if (res != null) {
       setState(() {
         _events = res.iruimList;
@@ -169,8 +177,8 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
       _loading = true;
     });
 
-    final res =
-        await GoPortApi.instance.getVehicleDetails(truck, driver.tz, trailer);
+    final res = await GoPortApi.instance
+        .getVehicleDetails(truck, driver!.tz ?? "", trailer);
 
     setState(() {
       _loading = false;
@@ -183,23 +191,18 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
       } else {
         generalProvider.setTruck(res.truck);
         generalProvider.setTrailer(res.trailer);
-
-        // generalProvider.truck = res.truck;
-        // generalProvider.trailer = res.trailer;
-
-        if (driver.companyNumber != generalProvider.truck.companyNumber ||
-            (trailer != null &&
-                driver.companyNumber !=
-                    generalProvider.trailer.companyNumber)) {
-          // generalProvider.truck = null;
-          // generalProvider.trailer = null;
+        if (driver.companyNumber != generalProvider.truck?.companyNumber) {
           generalProvider.setTruck(null);
           generalProvider.setTrailer(null);
           Utils.showAlertDialog(
               context: context,
               message: AppLocalizations.of(context)
                   .translate("Truck or Trailer Number Not Match"));
+          _controlsEnabled = false;
         } else {
+          // generalProvider.truck = res.truck;
+          // generalProvider.trailer = res.trailer;
+
           _initializeData();
           setState(() {
             _controlsEnabled = true;
@@ -272,14 +275,13 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
     });
 
     final generalProvider =
-    Provider.of<GeneralProvider>(context, listen: false);
+        Provider.of<GeneralProvider>(context, listen: false);
     final driver = generalProvider.driver;
-    final res = await GoPortApi.instance.getDriverGuidID(driver.tz);
+    final res = await GoPortApi.instance.getDriverGuidID(driver!.tz ?? "");
 
     setState(() {
       _loading = false;
     });
-
 
     if (res != null && res != "null") {
       generalProvider.setShowBackButton(true);
@@ -288,12 +290,14 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
       generalProvider.setShowBackButton(false);
       _checkJobCardExists();
     } else {
-      Utils.showAlertDialog(context: context, title: AppLocalizations.of(context)
-          .translate("You are not in the port area"));
+      Utils.showAlertDialog(
+          context: context,
+          title: AppLocalizations.of(context)
+              .translate("You are not in the port area"));
     }
   }
 
-  onContainersTapped() async{
+  onContainersTapped() async {
     if (_controlsEnabled) {
       final generalProvider =
           Provider.of<GeneralProvider>(context, listen: false);
@@ -446,7 +450,9 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
                                                 width: 40,
                                                 fit: BoxFit.contain,
                                               ),
-                                              SizedBox(height: 10,),
+                                              SizedBox(
+                                                height: 10,
+                                              ),
                                               Text(
                                                 AppLocalizations.of(context)
                                                     .translate("Containers"),
@@ -699,7 +705,7 @@ class _ActionTypeScreenState extends State<ActionTypeScreen>
                               ],
                             ),
                             visible: generalProvider.truck != null &&
-                                generalProvider.truck.isByPass)
+                                generalProvider.truck!.isByPass)
                       ],
                     )
                   ],
